Feature/review fixes (#2)
* Added backend & frontend tweaks based on review.

+++ src/BidCalculator/BidCalculator.Server/Core/SellerFee.cs
@@ -1,6 +1,6 @@
 ﻿namespace BidCalculator.Server.Core;
 
-public class SellerFee(double amount) : Fee(amount)
+public class SellerFee(decimal amount) : Fee(amount)
 {
-	public override string Name => "Seller";
+    public override string Name => "Seller";
 }
