+++ src/BidCalculator/BidCalculator.Server/Core/StorageFee.cs
@@ -1,6 +1,6 @@
 ﻿namespace BidCalculator.Server.Core;
 
-public class StorageFee(double amount) : Fee(amount)
+public class StorageFee(decimal amount) : Fee(amount)
 {
-	public override string Name => "Storage";
+    public override string Name => "Storage";
 }
